Found some logic errors in game board and update view
The app was crashing because there was incorrect logic in the game board
class that was making it impossible for the board to ever be full. I
also swapped around the views a little bit to make it reload, although I
intend to rip out the duplication of the new/play.

This is not part of my coding challenge answer; I just wanted to figure
it out for my own sake.